requirements.txt
requests~=2.24.0
beautifulsoup4~=4.9.1
discord.py